requirements.txt
aws-cdk.core==1.124.0

aws_cdk.aws-lambda==1.124.0
aws_cdk.aws_lambda_python==1.124.0
aws-cdk.aws_apigatewayv2==1.124.0
aws-cdk.aws_apigatewayv2_integrations==1.124.0
aws_cdk.aws_apigatewayv2_authorizers==1.124.0

pytest
